This PR makes some changes that lead to major performance gains for the surface forcing creation, in particular in terms of memory footprint. 

- [x] Parallelize computation of radiation correction (when `use_dask = True`), which leads to a huge improvement of the memory footprint, both during the validation step and the saving step (see #201). Closes #225.
- [x] For radiation correction, swap order of temporal and spatial interpolation: first temporal, then spatial. This somehow leads to a further improved memory footprint. E.g., performance for initialization of surface forcing with `correct_radiation = True`, `use_dask = True`, `bypass_validation = False`:
* first spatial, then temporal (old)
```
peak memory: 13589.34 MiB, increment: 13049.53 MiB
CPU times: user 20.2 s, sys: 3.85 s, total: 24 s
Wall time: 26.7 s
```
* first temporal, then spatial (new)
```
peak memory: 1343.37 MiB, increment: 803.54 MiB
CPU times: user 11 s, sys: 670 ms, total: 11.7 s
Wall time: 11.6 s
```
- [x] Fix bug in validation step: Always use rho-mask, which is the correct one.

Checklist:

- [x] Passes `pre-commit run --all-files`
- [x] Changes are documented in `docs/releases.md`